moved checkLogFiles to modules/utils
